Refactor: Consolidate kit commands and update UI color
The `addkit` command logic has been moved from its own file into `kit.js` to consolidate related functionality.

The "No, go back" button in the kit deletion confirmation dialog has been changed from light green to dark green for better UI consistency.
