+++ AddonExeBP/scripts/modules/commands/kit.js
@@ -3,6 +3,8 @@ import { commandManager } from './commandManager.js';
 import * as kitsManager from '../../core/kitsManager.js';
 import { getConfig } from '../../core/configManager.js';
 import { errorLog } from '../../core/errorLogger.js';
+import { createKit } from '../../core/kitAdminManager.js';
+import { addItemToKit } from '../../core/kitItemsManager.js';
 
 commandManager.register({
     name: 'kit',
@@ -57,5 +59,48 @@ commandManager.register({
         } else {
             player.sendMessage(`§c${result.message}`);
         }
+    }
+});
+
+commandManager.register({
+    name: 'addkit',
+    description: 'Create a new kit from your inventory.',
+    permissionLevel: 1, // Admins only
+    allowConsole: false,
+    parameters: [
+        { name: 'kitName', type: 'string', description: 'The name of the kit to create.' }
+    ],
+    execute: (player, args) => {
+        const { kitName } = args;
+
+        const inventory = player.getComponent('minecraft:inventory').container;
+        const items = [];
+        // Player inventory is slots 0-35.
+        for (let i = 0; i < 36; i++) {
+            const item = inventory.getItem(i);
+            if (item) {
+                items.push({
+                    typeId: item.typeId,
+                    amount: item.amount,
+                    nameTag: item.nameTag,
+                    lore: item.getLore()
+                });
+            }
+        }
+
+        if (items.length === 0) {
+            return player.sendMessage('§cYour inventory is empty. Cannot create an empty kit.');
+        }
+
+        const createResult = createKit(kitName);
+        if (!createResult.success) {
+            return player.sendMessage(`§c${createResult.message}`);
+        }
+
+        for (const item of items) {
+            addItemToKit(kitName, item);
+        }
+
+        player.sendMessage(`§aSuccessfully created kit '${kitName}' with ${items.length} item stacks.`);
     }
 });
